requirements/base.txt
Django==3.2.5
feedparser==6.0.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-nested-routers==0.93.3